fix: Refactor InnerTube data classes and update UI
- Moved all serializable data classes from `extractors.kt` to a new `Serializables.kt` file for better organization.
- Added `PlaylistScreen` to the list of screens where navigation elements should be hidden.
- Fixed a bug on the SearchScreen to prevent crashing when the search history contains an empty string.
- Updated the `PlayerScreen` slider colors to use `inversePrimary` and `secondary` from the MaterialTheme.

+++ app/src/main/kotlin/com/kynarec/kmusic/ui/screens/SearchScreen.kt
@@ -231,40 +231,44 @@ fun SearchScreen(
 //            horizontalAlignment = Alignment.CenterHorizontally
         ) {
             items(count = searchQueries.size) { index ->
-                Row(Modifier
-                    .fillMaxWidth()
-                    .padding(vertical = 4.dp)
-                    .clickable {
-                        keyboardController?.hide()
-                        focusManager.clearFocus()
-                        scope.launch {
-                            searchQueryDao.deleteQuery(searchQuery.text)
-                            searchQueryDao.insertQuery(SearchQuery(query = searchQuery.text))
-                        }
-                        navController.navigate(SearchResultScreen(searchQueries[index].query))
-                    },
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
+                if (searchQueries[index].query != "") {
+                    Row(
+                        Modifier
+                            .fillMaxWidth()
+                            .padding(vertical = 4.dp)
+                            .clickable {
+                                keyboardController?.hide()
+                                focusManager.clearFocus()
+                                scope.launch {
+                                    searchQueryDao.deleteQuery(searchQuery.text)
+                                    searchQueryDao.insertQuery(SearchQuery(query = searchQuery.text))
+                                }
+                                navController.navigate(SearchResultScreen(searchQueries[index].query))
+                            },
+                        verticalAlignment = Alignment.CenterVertically
+                    ) {
 
-                    Text(searchQueries[index].query)
+                        Text(searchQueries[index].query)
 
 
-                    Spacer(Modifier.weight(1f))
-                    IconButton(onClick = {
-                        scope.launch {
-                            searchQueryDao.deleteQuery(searchQueries[index].query)
+                        Spacer(Modifier.weight(1f))
+                        IconButton(onClick = {
+                            scope.launch {
+                                searchQueryDao.deleteQuery(searchQueries[index].query)
+                            }
+                        }) {
+                            Icon(imageVector = Icons.Default.Delete, contentDescription = "Delete")
                         }
-                    }) {
-                        Icon(imageVector = Icons.Default.Delete, contentDescription = "Delete")
-                    }
-                    IconButton(onClick = {
-                        scope.launch {
-                            searchQuery = TextFieldValue(searchQueries[index].query,
-                                selection = TextRange(searchQueries[index].query.length)
-                            )
+                        IconButton(onClick = {
+                            scope.launch {
+                                searchQuery = TextFieldValue(
+                                    searchQueries[index].query,
+                                    selection = TextRange(searchQueries[index].query.length)
+                                )
+                            }
+                        }) {
+                            Icon(imageVector = Icons.Default.Edit, contentDescription = "Edit")
                         }
-                    }) {
-                        Icon(imageVector = Icons.Default.Edit, contentDescription = "Edit")
                     }
                 }
             }
